Fix pipeline front end stalling
Sources signal for the (now more appropriately-named) front end enable
OR gate from do_branch, rather than the output of the EX/MEM clear OR
gate. The 3-slot predict-not-taken model now passes the co-simulation
test, though not the main unit test for some reason.

Also updates the layouts for this model.