right i was making it do an infinite loop
The state of the form is now polled periodically instead of when an
option is changed, which is what actually caused the crash from the
previous commit. Processor variants do not currently update correctly,
limiting the amount of processors that are possible to select, but the
dialog otherwise works and is able to set the selected processor.